+++ CHANGELOG.md
@@ -3,12 +3,19 @@ All notable changes to this project will be documented in this file.
 
 ## v0.2.4-alpha - 2024-04-23
 
+
 ### Changed
 - Hooks are registered as structs with an ID so they are easier to add/remove at runtime.
 - Renamed `HookManager.register_hook_handlers/3` to `HookManager.register_handlers/3`.
 
 ### Added
 - Function to remove hook handler by it's ID. `HookManager.cancel_handler/3`.
+- `ProcessHub.Strategy.PartitionTolerance.Base.toggle_lock?/3`
+- `ProcessHub.Strategy.PartitionTolerance.Base.toggle_unlock?/3`
+
+### Removed
+- `ProcessHub.Strategy.PartitionTolerance.Base.handle_node_up/3`
+- `ProcessHub.Strategy.PartitionTolerance.Base.handle_node_down/3`
 
 ## v0.2.3-alpha - 2024-04-23
 Added new configuration options for timeout values and renamed LocalStorage -> Storage.